+++ packages/active_grid_core/lib/active_grid_model.dart
@@ -14,10 +14,15 @@ part 'model/form/component/integer_form_component.dart';
 part 'model/form/component/date_form_component.dart';
 part 'model/form/component/enum_form_component.dart';
 part 'model/form/component/options.dart';
-part 'model/form/form_data.dart';
 part 'model/form/form_actions.dart';
+part 'model/form/form_data.dart';
+part 'model/form/form_uri.dart';
 
 part 'model/grid/grid.dart';
 part 'model/grid/grid_field.dart';
 part 'model/grid/grid_entry.dart';
 part 'model/grid/grid_row.dart';
+
+part 'model/space/space.dart';
+
+part 'model/user/user.dart';