+++ packages/active_grid_core/test/active_grid_core_test.dart
@@ -6,7 +6,6 @@ import 'package:provider/provider.dart';
 import 'mocks.dart';
 
 void main() {
-
   group('ActiveGrid Provides necessary Objects', () {
     testWidgets('ActiveGridClient', (tester) async {
       late BuildContext context;