+++ packages/active_grid_form/CHANGELOG.md
@@ -1,2 +1,6 @@
+## [0.2.0]
+* Load Forms using FormUri
+* Supports accessing Forms with a direct Uri representation
+
 ## [0.1.0]
 * Initial Release